Consumer inserts the events to database
This does not do any math here, the DB is using Timescale
hypertable that is great for this kind of timeseries data.
This however requires that the source events are kept, leading to
a ton of data in the DB. The retention policy should be set to
clear the data from the DB, potentially pushing the data to some
cold storage (AWS S3 Glacier?) if the source data is important.

The summary data can be kept in a materialized views and calculated
automatically.